+++ packages/cordova-plugin/vite.location-button.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  build: {
+    emptyOutDir: false,
+    minify: false,
+    outDir: 'dist',
+    target: 'es2020',
+    lib: {
+      entry: './src/location-button.ts',
+      name: 'OSGeolocationLocationButton',
+      fileName: () => 'location-button.js',
+      formats: ['iife'],
+    },
+  },
+});